BUGFIX: Finally fixed close flashing issue.
It must be called BEFORE WM_DESTROY because WM_DESTROY is emitted 
AFTER the childs and itself have been destroyed. In the resulting, 
short time interval (after destroy and before handling WM_DESTROY)
a new flashing can occur, because the Window Messaging System
seems to give the keyboard/mouse input to the next window. But
without waiting for the WM_DESTROY to be processed. That's why
we're directly doing it in WM_CLOSE now :-)