Fix #298: APE decoding fails
This was a more comprehensive set of changes than a simple bug fix. The original issue highlighted a pretty nasty issue in the way we were dealing with multi-frame packets. The entire plugin was refactored, and now uses modern, non-deprecated APIs like `avcodec_send_packet` and `avcodec_receive_frame`, and also utilizes `AVAudioFifo` and `av_audio_fifo_*` APIs to ensure stable buffer sizes are returned to the caller.